EPMRPP-49009: Move NoCasesBlock to common component and expose it to plugins.
