+++ app/src/components/main/noCasesBlock/index.js
