+++ app/src/components/main/noCasesBlock/noCasesBlock.jsx
