+++ app/src/components/main/noCasesBlock/noCasesBlock.scss
