+++ app/src/pages/common/settingsPage/patternAnalysisTab/patternAnalysisTab.jsx
@@ -35,9 +35,9 @@ import { GhostButton } from 'components/buttons/ghostButton';
 import PlusIcon from 'common/img/plus-button-inline.svg';
 import { canUpdateSettings } from 'common/utils/permissions';
 import { activeProjectRoleSelector, userAccountRoleSelector } from 'controllers/user';
+import { NoCasesBlock } from 'components/main/noCasesBlock';
 import { RuleListHeader } from '../ruleListHeader';
 import { RuleList } from '../ruleList';
-import { NoCasesBlock } from '../noCasesBlock';
 import styles from './patternAnalysisTab.scss';
 import { messages } from './messages';
